Move ImageManipulation to own file
